+++ tests/pallas/tpu_pallas_distributed_test.py
@@ -56,12 +56,12 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
     def kernel(x_ref, y_ref):
       def body(ready_sem, send_sem, recv_sem):
         other_dev_id = 1 - lax.axis_index('x')
-        pltpu.semaphore_signal(ready_sem, device_id=other_dev_id,
-                               device_id_type=pltpu.DeviceIdType.LOGICAL)
-        pltpu.semaphore_wait(ready_sem)
+        pl.semaphore_signal(ready_sem, device_id=other_dev_id,
+                               device_id_type=pl.DeviceIdType.LOGICAL)
+        pl.semaphore_wait(ready_sem)
         copy_done = pltpu.async_remote_copy(
             x_ref, y_ref, send_sem, recv_sem, other_dev_id,
-            device_id_type=pltpu.DeviceIdType.LOGICAL,
+            device_id_type=pl.DeviceIdType.LOGICAL,
         )
         copy_done.wait_send()
         copy_done.wait_recv()
@@ -104,12 +104,12 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
     def kernel(x_ref, y_ref):
       def body(ready_sem, send_sem, recv_sem):
         other_dev_id = 1 - lax.axis_index('x')
-        pltpu.semaphore_signal(ready_sem, device_id=other_dev_id,
-                               device_id_type=pltpu.DeviceIdType.LOGICAL)
-        pltpu.semaphore_wait(ready_sem)
+        pl.semaphore_signal(ready_sem, device_id=other_dev_id,
+                               device_id_type=pl.DeviceIdType.LOGICAL)
+        pl.semaphore_wait(ready_sem)
         copy_done = pltpu.async_remote_copy(
             x_ref, y_ref, send_sem, recv_sem, other_dev_id,
-            device_id_type=pltpu.DeviceIdType.LOGICAL,
+            device_id_type=pl.DeviceIdType.LOGICAL,
         )
         copy_done.wait_send()
         copy_done.wait_recv()
@@ -159,8 +159,8 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
     def kernel(x_ref, y_ref):
       def body(ready_sem, send_sem, recv_sem):
         other_dev_id = 1 - lax.axis_index('x')
-        pltpu.semaphore_signal(ready_sem, device_id={'x': other_dev_id})
-        pltpu.semaphore_wait(ready_sem)
+        pl.semaphore_signal(ready_sem, device_id={'x': other_dev_id})
+        pl.semaphore_wait(ready_sem)
         pltpu.async_remote_copy(
             x_ref,
             y_ref,
@@ -168,8 +168,8 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
             recv_sem,
             device_id={'x': other_dev_id},
         )
-        pltpu.semaphore_wait(send_sem, 1)
-        pltpu.semaphore_wait(recv_sem, 1)
+        pl.semaphore_wait(send_sem, 1)
+        pl.semaphore_wait(recv_sem, 1)
 
       pl.run_scoped(
           body,
@@ -220,8 +220,8 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
           neighbor = lax.rem(my_id - 1, num_devices)
           # Neighbor might be negative here so we add num_devices in case
           neighbor = jnp.where(neighbor < 0, neighbor + num_devices, neighbor)
-        pltpu.semaphore_signal(ready_sem, device_id=neighbor)
-        pltpu.semaphore_wait(ready_sem)
+        pl.semaphore_signal(ready_sem, device_id=neighbor)
+        pl.semaphore_wait(ready_sem)
         copy_done = pltpu.async_remote_copy(
             x_ref, y_ref, send_sem, recv_sem, device_id=neighbor
         )
@@ -274,8 +274,8 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
           neighbor = lax.rem(my_id - 1, axis_size)
           # Neighbor might be negative here so we add num_devices in case
           neighbor = jnp.where(neighbor < 0, neighbor + axis_size, neighbor)
-        pltpu.semaphore_signal(ready_sem, device_id=(my_other_id, neighbor))
-        pltpu.semaphore_wait(ready_sem)
+        pl.semaphore_signal(ready_sem, device_id=(my_other_id, neighbor))
+        pl.semaphore_wait(ready_sem)
         copy_done = pltpu.async_remote_copy(
             x_ref, y_ref, send_sem, recv_sem, device_id=(my_other_id, neighbor)
         )
@@ -326,10 +326,10 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
         num_devices = lax.axis_size('x')
         neighbor = lax.rem(my_id + 1, num_devices)
         barrier_sem = pltpu.get_barrier_semaphore()
-        pltpu.semaphore_signal(barrier_sem, device_id=neighbor)
-        pltpu.semaphore_wait(barrier_sem)
-        pltpu.semaphore_signal(ready_sem, device_id=neighbor)
-        pltpu.semaphore_wait(ready_sem)
+        pl.semaphore_signal(barrier_sem, device_id=neighbor)
+        pl.semaphore_wait(barrier_sem)
+        pl.semaphore_signal(ready_sem, device_id=neighbor)
+        pl.semaphore_wait(ready_sem)
         pltpu.async_remote_copy(
             x_ref, y_ref, send_sem, recv_sem, device_id=neighbor
         ).wait()
@@ -369,8 +369,8 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
       num_devices = lax.axis_size('x')
       barrier_sem = pltpu.get_barrier_semaphore()
       for i in range(num_devices):
-        pltpu.semaphore_signal(barrier_sem, device_id=i)
-      pltpu.semaphore_wait(barrier_sem, num_devices)
+        pl.semaphore_signal(barrier_sem, device_id=i)
+      pl.semaphore_wait(barrier_sem, num_devices)
       pltpu.sync_copy(x_ref, y_ref)
 
     x = jnp.arange(8 * 128).reshape((8, 128))
@@ -433,8 +433,8 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
           core_index = jax.lax.axis_index('core')
           # Make sure all cores have entered run_scoped.
           for j in range(num_cores):
-            pltpu.semaphore_signal(core_sem, 1, device_id={'core': j})
-          pltpu.semaphore_wait(core_sem, num_cores)
+            pl.semaphore_signal(core_sem, 1, device_id={'core': j})
+          pl.semaphore_wait(core_sem, num_cores)
 
           device_index = jax.lax.axis_index('device')
           slc = pl.ds(core_index * slc_size, slc_size)
@@ -443,10 +443,10 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
           sem0 = pltpu.get_barrier_semaphore()
           for i in range(ddim):
             for j in range(num_cores):
-              pltpu.semaphore_signal(
+              pl.semaphore_signal(
                   sem0, 1, device_id={'device': i, 'core': j}
               )
-          pltpu.semaphore_wait(sem0, ddim * num_cores)
+          pl.semaphore_wait(sem0, ddim * num_cores)
 
           # Identity function by default
           pltpu.async_copy(in_ref.at[:, slc], out_ref.at[:, slc], sem).wait()
@@ -461,7 +461,7 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
               send_sem=send_sem,
               recv_sem=recv_sem,
               device_id=device_id,
-              device_id_type=pltpu.DeviceIdType.MESH,
+              device_id_type=pl.DeviceIdType.MESH,
           )
 
           @pl.when(device_index == 0)
@@ -518,8 +518,8 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
         num_cores = jax.lax.axis_size('core')
         def inner(sem):
           for i in range(num_cores):
-            pltpu.semaphore_signal(sem, 1, device_id={'core': i})
-          pltpu.semaphore_wait(sem, num_cores)
+            pl.semaphore_signal(sem, 1, device_id={'core': i})
+          pl.semaphore_wait(sem, num_cores)
           core_id = jax.lax.axis_index('core')
           pltpu.sync_copy(x_ref.at[:, core_id], y_ref.at[:, core_id])
         pl.run_scoped(inner, pltpu.SemaphoreType.REGULAR)
@@ -537,14 +537,14 @@ class PallasCallRemoteDMATest(parameterized.TestCase):
       num_devices = lax.axis_size('x')
       barrier_sem = pltpu.get_barrier_semaphore()
       for i in range(num_devices):
-        pltpu.semaphore_signal(barrier_sem, device_id=i)
-      pltpu.semaphore_wait(barrier_sem, num_devices)
+        pl.semaphore_signal(barrier_sem, device_id=i)
+      pl.semaphore_wait(barrier_sem, num_devices)
 
     def barrier_kernel(x_ref, sem_ref, out_ref):
       num_devices = lax.axis_size('x')
       for i in range(num_devices):
-        pltpu.semaphore_signal(sem_ref, device_id=i)
-      pltpu.semaphore_wait(sem_ref, num_devices)
+        pl.semaphore_signal(sem_ref, device_id=i)
+      pl.semaphore_wait(sem_ref, num_devices)
       out_ref[...] = x_ref[...] + 1
 
     x = jnp.arange(8 * 128).reshape((8, 128))
@@ -614,7 +614,7 @@ class PallasCallRemoteDMAInterpretTest(parameterized.TestCase):
           send_sem=copy_send_sem,
           recv_sem=copy_recv_sem,
           device_id=dst_device,
-          device_id_type=pltpu.DeviceIdType.LOGICAL,
+          device_id_type=pl.DeviceIdType.LOGICAL,
       )
       input_to_output_copy.start()
       input_to_output_copy.wait()
@@ -686,8 +686,8 @@ class PallasCallRemoteDMAInterpretTest(parameterized.TestCase):
       # Signal to the sender to this device that output_ref has been zeroed
       # and this device is ready to receive.
       # prev_device = (my_id - 1) % num_devices
-      # pltpu.semaphore_signal(barrier_sem, 1, device_id=prev_device)
-      # pltpu.semaphore_wait(barrier_sem)
+      # pl.semaphore_signal(barrier_sem, 1, device_id=prev_device)
+      # pl.semaphore_wait(barrier_sem)
 
       # If the device_id is even, we copy to output_ref[1].
       # If it's odd, we copy to output_ref[0].
@@ -790,7 +790,7 @@ class PallasCallRemoteDMAInterpretTest(parameterized.TestCase):
             send_sem=send_sem,
             recv_sem=recv_sem,
             device_id=neighbor,
-            device_id_type=pltpu.DeviceIdType.LOGICAL,
+            device_id_type=pl.DeviceIdType.LOGICAL,
         )
         remote_dma.start()
         remote_dma.wait()
@@ -802,7 +802,7 @@ class PallasCallRemoteDMAInterpretTest(parameterized.TestCase):
             send_sem=send_sem,
             recv_sem=recv_sem,
             device_id=neighbor,
-            device_id_type=pltpu.DeviceIdType.LOGICAL,
+            device_id_type=pl.DeviceIdType.LOGICAL,
         )
         remote_dma.start()
         remote_dma.wait()
@@ -880,7 +880,7 @@ class VerificationTest(jtu.JaxTestCase):
       my_id = lax.axis_index('x')
       dst_id = jnp.where(my_id == num_devices - 1, 0, my_id + 1)
       src_id = jnp.where(my_id == 0, num_devices - 1, my_id - 1)
-      pltpu.semaphore_signal(capacity_sem, 1, device_id=src_id)
+      pl.semaphore_signal(capacity_sem, 1, device_id=src_id)
       out_ref[...] = jnp.zeros_like(out_ref)
       scratch_ref[0] = in_ref[0]
 
@@ -888,7 +888,7 @@ class VerificationTest(jtu.JaxTestCase):
       def _(i, _):
         slot = i % 2
         next_slot = 1 - slot
-        pltpu.semaphore_wait(capacity_sem, 1)
+        pl.semaphore_wait(capacity_sem, 1)
         copy = pltpu.async_remote_copy(
             scratch_ref.at[slot],
             scratch_ref.at[next_slot],
@@ -898,9 +898,9 @@ class VerificationTest(jtu.JaxTestCase):
         )
         out_ref[...] += scratch_ref[slot]
         copy.wait()
-        pltpu.semaphore_signal(capacity_sem, 1, device_id=src_id)
+        pl.semaphore_signal(capacity_sem, 1, device_id=src_id)
       out_ref[...] += scratch_ref[(num_devices - 1) % 2]
-      pltpu.semaphore_wait(capacity_sem, 1)
+      pl.semaphore_wait(capacity_sem, 1)
 
     kernel = pl.pallas_call(
         kernel_body,